Move SceneDocument into backend/domain/graph.py (slice 2)
Second slice of ADR-002 stage 2.2. The entire SceneDocument dataclass
(through grid_payload) plus the _content_parts_wire helper scene_payload
calls moved VERBATIM to backend/domain/graph.py, with exactly ONE
deliberate edit, diffed and proven to be the only changed line:
add_session_tokens now calls a local _estimate_tokens helper instead of
backend.token_counter.estimate_tokens, because token_counter imports
backend.events at module level, which the domain purity gate forbids.
The helper replicates that function verbatim (TokenEstimator().
count_tokens(text), same fresh-instance-per-call shape), and the call
site's own str() coercion is unchanged, so token counts - which feed
total_session_tokens, saved chats.db rows, and autosave's change hash -
are bit-identical.

render_chart_png stays a domain import deliberately: a pure
(type, data, w, h, scale) -> PNG-bytes function, and the chart node's
versioned rendered asset is a genuine document invariant - the same
root-module dependency category as NavigationPinStore/GridViewSettings,
which the class has always carried as fields. Module-level import also
preserves today's matplotlib import-order side effects.

canvas.py (5,084 lines at the start of this stage) is now 1,810 lines -
the orchestration/wire/registration layer only - already under the
stage's 2,500-line exit criterion. Its imports were trimmed to what the
remainder actually uses (verified per name: itertools/math/uuid/
dataclasses/logging/render_chart_png/estimate_tokens/GridViewSettings/
NavigationPinStore/GRID_SIZE_PRESETS all had zero non-comment uses
left); the full external compat surface - every name session_load/
session_save/plugins/app/tests import from backend.canvas, including
the private _content_codec (same shared instance, identity asserted) -
is preserved and import-tested.

Verified: dataclass signatures/field order/defaults for SceneDocument/
SceneNode/SceneEdge diffed against the pre-refactor baseline
(identical, address-normalized); full suite 1,187 passed.
